+++ packages/web-extension/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rrweb/web-extension
 
+## 2.0.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @saola.ai/rrweb@2.0.8
+  - @saola.ai/rrweb-player@2.0.8
+
 ## 2.0.7
 
 ### Patch Changes